+++ lib/http_name_inserter_test.go
@@ -67,9 +67,12 @@ func TestHTTPNameInserter(t *testing.T) {
 	}
 	err = hni.Insert(
 		SourceID{Location: SourceLocation{Repo: "a-repo", Dir: "offset"}, Version: semv.MustParse("5.5.5")},
-		"dockerthin.com/repo/latest",
-		"",
-		[]Quality{},
+		BuildArtifact{
+			DigestReference: "dockerthin.com/repo@sha256:012345678901234567890123456789AB012345678901234567890123456789AB",
+			//Name:      "dockerthin.com/repo/latest",
+			Type:      "docker",
+			Qualities: []Quality{},
+		},
 	)
 	if err != nil {
 		t.Error(err)